refactor: benchmark
