+++ package.json
@@ -50,7 +50,7 @@
     "fix:code": "yarn lint:code --fix",
     "fix:special": "node node_modules/tooling/inherit-types --write && node node_modules/tooling/format-schemas --write && node tooling/generate-runtime-code.js --write && node tooling/generate-wasm-code.js --write && node node_modules/tooling/compile-to-definitions --write && node node_modules/tooling/precompile-schemas --write && node node_modules/tooling/generate-types --no-template-literals --write",
     "build:examples": "cd examples && node buildAll.js",
-    "benchmark": "node --max-old-space-size=6144 --experimental-vm-modules --trace-deprecation --hash-seed=1 --random-seed=1 --no-opt --predictable --predictable-gc-schedule --interpreted-frames-native-stack --allow-natives-syntax --expose-gc --no-concurrent-sweeping ./test/BenchmarkTestCases.benchmark.mjs",
+    "benchmark": "node --max-old-space-size=4096 --experimental-vm-modules --expose-gc ./test/BenchmarkTestCases.benchmark.mjs",
     "pretest": "yarn lint",
     "test": "yarn test:base",
     "test:base": "node --expose-gc --max-old-space-size=4096 --experimental-vm-modules --trace-deprecation node_modules/jest-cli/bin/jest --logHeapUsage",
@@ -152,6 +152,7 @@
     "jest-diff": "^30.2.0",
     "jest-environment-node": "^30.2.0",
     "jest-junit": "^16.0.0",
+    "jest-worker": "^30.2.0",
     "json-loader": "^0.5.7",
     "json5": "^2.1.3",
     "less": "^4.5.1",
